refactor: move group.ts code into meeting/
